Moved out non-strategic publication integration tests (#676)
* Moved out non-strategic publication integration tests

* Rearranged fields

* Added integration test base to non-strategic publication tests
